% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-methods.R
\name{bind_keyed}
\alias{bind_keyed}
\title{Bind rows of keyed data frames}
\usage{
bind_keyed(..., .id = NULL)
}
\arguments{
\item{...}{Data frames to bind.}

\item{.id}{Optional column name to identify source.}
}
\value{
A keyed data frame if key is preserved and unique, otherwise tibble.
}
\description{
Bind keyed data frames
}
\details{
Wrapper for \code{\link[dplyr:bind_rows]{dplyr::bind_rows()}} that attempts to preserve key metadata.
}
