% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-methods.R
\name{bind_rows_keyed}
\alias{bind_rows_keyed}
\title{Bind rows of keyed data frames}
\usage{
bind_rows_keyed(..., .id = NULL)
}
\arguments{
\item{...}{Data frames to bind.}

\item{.id}{Optional column name to identify source.}
}
\value{
A keyed data frame if key is preserved and unique, otherwise tibble.
}
\description{
Wrapper for dplyr::bind_rows that attempts to preserve key.
}
