% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{clear_all_snapshots}
\alias{clear_all_snapshots}
\title{Clear all snapshots from cache}
\usage{
clear_all_snapshots(confirm = TRUE)
}
\arguments{
\item{confirm}{If TRUE, require confirmation.}
}
\description{
Clear all snapshots from cache
}
