% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{clear_snapshot}
\alias{clear_snapshot}
\title{Clear snapshot for a data frame}
\usage{
clear_snapshot(.data, purge = FALSE)
}
\arguments{
\item{.data}{A data frame.}

\item{purge}{If TRUE, also remove the snapshot from cache.}
}
\value{
Data frame without snapshot reference.
}
\description{
Removes the snapshot reference from a data frame.
}
