% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-id.R
\name{get_id}
\alias{get_id}
\title{Get ID column}
\usage{
get_id(.data, .id = ".id")
}
\arguments{
\item{.data}{A data frame.}

\item{.id}{Column name (default: ".id").}
}
\value{
Character vector of IDs, or NULL if not present.
}
\description{
Get ID column
}
