% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{get_key_cols}
\alias{get_key_cols}
\title{Get key column names}
\usage{
get_key_cols(.data)
}
\arguments{
\item{.data}{A keyed data frame.}
}
\value{
Character vector of column names, or NULL if no key.
}
\description{
Get key column names
}
