% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-id.R
\name{get_row_id}
\alias{get_row_id}
\title{Get row ID column}
\usage{
get_row_id(.data, .id = ".row_id")
}
\arguments{
\item{.data}{A data frame.}

\item{.id}{Column name (default: ".row_id").}
}
\value{
Character vector of row IDs, or NULL if not present.
}
\description{
Get row ID column
}
