% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-id.R
\name{has_id}
\alias{has_id}
\title{Check if data frame has IDs}
\usage{
has_id(.data, .id = ".id")
}
\arguments{
\item{.data}{A data frame.}

\item{.id}{Column name to check (default: ".id").}
}
\value{
Logical.
}
\description{
Check if data frame has IDs
}
