% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-id.R
\name{has_row_id}
\alias{has_row_id}
\title{Check if data frame has row IDs}
\usage{
has_row_id(.data, .id = ".row_id")
}
\arguments{
\item{.data}{A data frame.}

\item{.id}{Column name to check (default: ".row_id").}
}
\value{
Logical.
}
\description{
Check if data frame has row IDs
}
