% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key_is_valid}
\alias{key_is_valid}
\title{Check if the key is still valid}
\usage{
key_is_valid(.data)
}
\arguments{
\item{.data}{A keyed data frame.}
}
\value{
Logical. Returns FALSE with a warning if key is invalid.
}
\description{
Checks whether the key columns still exist and are still unique.
}
