% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{list_snapshots}
\alias{list_snapshots}
\title{List all snapshots in cache}
\usage{
list_snapshots()
}
\value{
Data frame with snapshot information, including \code{size_mb}.
}
\description{
List all snapshots in cache
}
