% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-id.R
\name{remove_id}
\alias{remove_id}
\title{Remove ID column}
\usage{
remove_id(.data, .id = ".id")
}
\arguments{
\item{.data}{A data frame.}

\item{.id}{Column name to remove (default: ".id").}
}
\value{
Data frame without the ID column.
}
\description{
Remove ID column
}
