% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-id.R
\name{remove_row_id}
\alias{remove_row_id}
\title{Remove row ID column}
\usage{
remove_row_id(.data, .id = ".row_id")
}
\arguments{
\item{.data}{A data frame.}

\item{.id}{Column name to remove (default: ".row_id").}
}
\value{
Data frame without the row ID column.
}
\description{
Remove row ID column
}
