% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{summary.keyed_df}
\alias{summary.keyed_df}
\title{Summary method for keyed data frames}
\usage{
\method{summary}{keyed_df}(object, ...)
}
\arguments{
\item{object}{A keyed data frame.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns a summary list.
}
\description{
Summary method for keyed data frames
}
