% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{unkey}
\alias{unkey}
\title{Remove key from a data frame}
\usage{
unkey(.data)
}
\arguments{
\item{.data}{A keyed data frame.}
}
\value{
A tibble without key metadata.
}
\description{
Remove key from a data frame
}
